Refactor so everything is inside App
